refactor(sim): move control ABI to hardware boundary
