+++ dimos/hardware/manipulators/sim/adapter.py
@@ -27,7 +27,7 @@ from dimos.hardware.manipulators.spec import (
     JointLimits,
     ManipulatorInfo,
 )
-from dimos.simulation.engines.mujoco_shm import (
+from dimos.hardware.simulation.shared_memory import (
     ManipShmReader,
     shm_key_from_path,
 )
